+++ tests/conftest.py
@@ -209,7 +209,7 @@ def meson_fatal_warnings():
     mpatch = pytest.MonkeyPatch()
     mesonpy_project_init = mesonpy.Project.__init__
 
-    def __init__(self, source_dir, build_dir, meson_args=None, editable_verbose=None):
+    def __init__(self, source_dir, build_dir, meson_args=None, editable_verbose=None, *args, **kwargs):
         if pathlib.Path(source_dir).absolute().name not in {
 
                 # The CMake subproject emits ``WARNING: CMake Toolchain:
@@ -228,6 +228,6 @@ def meson_fatal_warnings():
             if meson_args is None:
                 meson_args = {}
             meson_args.setdefault('setup', []).append('--fatal-meson-warnings')
-        mesonpy_project_init(self, source_dir, build_dir, meson_args, editable_verbose)
+        mesonpy_project_init(self, source_dir, build_dir, meson_args, editable_verbose, *args, **kwargs)
 
     mpatch.setattr(mesonpy.Project, '__init__', __init__)